requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.8.2
beautifulsoup4==4.12.3
python-multipart==0.0.9
PyYAML==6.0.2
